+++ survey3/src/components/BackButton.vue
@@ -1,5 +1,6 @@
 <template>
-  <div class="absolute-top-right q-mr-sm" :class="margintop">
+  <!-- QUEST MODE: hidden behind 3-dot menu -->
+  <div v-if="hidden" class="absolute-top-right q-mr-sm" :class="margintop">
     <q-btn color="grey-7" round flat icon="more_vert" data-cy="btn_options">
       <q-menu cover auto-close>
         <q-list>
@@ -10,6 +11,21 @@
       </q-menu>
     </q-btn>
   </div>
+
+  <!-- DEFAULT: visible back button -->
+  <div v-else class="absolute-top-left q-ml-sm" :class="margintop">
+    <q-btn
+      flat
+      round
+      dense
+      icon="arrow_back"
+      color="grey-8"
+      data-cy="back_root"
+      @click="quitForm"
+    >
+      <q-tooltip>{{$t('btn.back.label')}}</q-tooltip>
+    </q-btn>
+  </div>
 </template>
 
 
@@ -17,7 +33,7 @@
 
 export default {
   name: 'BACKBUTTON',
-  props: ["ask", "go_back", "go_location"],
+  props: ["ask", "go_back", "go_location", "hidden"],
 
   data() {
     return {}
@@ -25,7 +41,6 @@ export default {
 
   computed: {
     margintop() {
-      // console.log(this.$q.platform.is)
       if (this.$q.platform.is.iphone) return 'q-mt-xl'
       if (this.$q.platform.is.cordova) return 'q-mt-lg'
       return 'q-mt-md'
@@ -33,7 +48,7 @@ export default {
   },
 
   methods: {
-  
+
     quitForm() {
       if (this.go_location !== undefined) this.$router.push(this.go_location)
       if (this.go_back === true) return this.$router.go(-1)